requirements.txt
line-bot-sdk==3.5.1
flask==3.0.0
APScheduler==3.10.4
gunicorn==21.2.0
google-generativeai==0.3.2
